# ai-boilerplate-backend/api/admin/v1/membership_benefit_type_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.5.3
// - protoc             v3.21.9
// source: admin/v1/membership_benefit_type.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMembershipBenefitTypeCreateMembershipBenefitType = "/admin.v1.MembershipBenefitType/CreateMembershipBenefitType"
const OperationMembershipBenefitTypeDeleteMembershipBenefitType = "/admin.v1.MembershipBenefitType/DeleteMembershipBenefitType"
const OperationMembershipBenefitTypeGetMembershipBenefitTypeInfo = "/admin.v1.MembershipBenefitType/GetMembershipBenefitTypeInfo"
const OperationMembershipBenefitTypeGetMembershipBenefitTypeList = "/admin.v1.MembershipBenefitType/GetMembershipBenefitTypeList"
const OperationMembershipBenefitTypeUpdateMembershipBenefitType = "/admin.v1.MembershipBenefitType/UpdateMembershipBenefitType"
const OperationMembershipBenefitTypeUpdateMembershipBenefitTypeStatus = "/admin.v1.MembershipBenefitType/UpdateMembershipBenefitTypeStatus"

type MembershipBenefitTypeHTTPServer interface {
	CreateMembershipBenefitType(context.Context, *CreateMembershipBenefitTypeReq) (*CreateMembershipBenefitTypeReply, error)
	DeleteMembershipBenefitType(context.Context, *DeleteMembershipBenefitTypeReq) (*DeleteMembershipBenefitTypeReply, error)
	GetMembershipBenefitTypeInfo(context.Context, *GetMembershipBenefitTypeInfoReq) (*GetMembershipBenefitTypeInfoReply, error)
	GetMembershipBenefitTypeList(context.Context, *GetMembershipBenefitTypeListReq) (*GetMembershipBenefitTypeListReply, error)
	UpdateMembershipBenefitType(context.Context, *UpdateMembershipBenefitTypeReq) (*UpdateMembershipBenefitTypeReply, error)
	UpdateMembershipBenefitTypeStatus(context.Context, *UpdateMembershipBenefitTypeStatusReq) (*UpdateMembershipBenefitTypeStatusReply, error)
}

func RegisterMembershipBenefitTypeHTTPServer(s *http.Server, srv MembershipBenefitTypeHTTPServer) {
	r := s.Route("/")
	r.POST("/admin/v1/membership_benefit_type/create", _MembershipBenefitType_CreateMembershipBenefitType0_HTTP_Handler(srv))
	r.POST("/admin/v1/membership_benefit_type/update", _MembershipBenefitType_UpdateMembershipBenefitType0_HTTP_Handler(srv))
	r.POST("/admin/v1/membership_benefit_type/update/status", _MembershipBenefitType_UpdateMembershipBenefitTypeStatus0_HTTP_Handler(srv))
	r.POST("/admin/v1/membership_benefit_type/delete", _MembershipBenefitType_DeleteMembershipBenefitType0_HTTP_Handler(srv))
	r.GET("/admin/v1/membership_benefit_type/info", _MembershipBenefitType_GetMembershipBenefitTypeInfo0_HTTP_Handler(srv))
	r.GET("/admin/v1/membership_benefit_type/list", _MembershipBenefitType_GetMembershipBenefitTypeList0_HTTP_Handler(srv))
}

func _MembershipBenefitType_CreateMembershipBenefitType0_HTTP_Handler(srv MembershipBenefitTypeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateMembershipBenefitTypeReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipBenefitTypeCreateMembershipBenefitType)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateMembershipBenefitType(ctx, req.(*CreateMembershipBenefitTypeReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateMembershipBenefitTypeReply)
		return ctx.Result(200, reply)
	}
}

func _MembershipBenefitType_UpdateMembershipBenefitType0_HTTP_Handler(srv MembershipBenefitTypeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMembershipBenefitTypeReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipBenefitTypeUpdateMembershipBenefitType)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMembershipBenefitType(ctx, req.(*UpdateMembershipBenefitTypeReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMembershipBenefitTypeReply)
		return ctx.Result(200, reply)
	}
}

func _MembershipBenefitType_UpdateMembershipBenefitTypeStatus0_HTTP_Handler(srv MembershipBenefitTypeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMembershipBenefitTypeStatusReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipBenefitTypeUpdateMembershipBenefitTypeStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMembershipBenefitTypeStatus(ctx, req.(*UpdateMembershipBenefitTypeStatusReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMembershipBenefitTypeStatusReply)
		return ctx.Result(200, reply)
	}
}

func _MembershipBenefitType_DeleteMembershipBenefitType0_HTTP_Handler(srv MembershipBenefitTypeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteMembershipBenefitTypeReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipBenefitTypeDeleteMembershipBenefitType)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteMembershipBenefitType(ctx, req.(*DeleteMembershipBenefitTypeReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteMembershipBenefitTypeReply)
		return ctx.Result(200, reply)
	}
}

func _MembershipBenefitType_GetMembershipBenefitTypeInfo0_HTTP_Handler(srv MembershipBenefitTypeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMembershipBenefitTypeInfoReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipBenefitTypeGetMembershipBenefitTypeInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMembershipBenefitTypeInfo(ctx, req.(*GetMembershipBenefitTypeInfoReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMembershipBenefitTypeInfoReply)
		return ctx.Result(200, reply)
	}
}

func _MembershipBenefitType_GetMembershipBenefitTypeList0_HTTP_Handler(srv MembershipBenefitTypeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMembershipBenefitTypeListReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMembershipBenefitTypeGetMembershipBenefitTypeList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMembershipBenefitTypeList(ctx, req.(*GetMembershipBenefitTypeListReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMembershipBenefitTypeListReply)
		return ctx.Result(200, reply)
	}
}

type MembershipBenefitTypeHTTPClient interface {
	CreateMembershipBenefitType(ctx context.Context, req *CreateMembershipBenefitTypeReq, opts ...http.CallOption) (rsp *CreateMembershipBenefitTypeReply, err error)
	DeleteMembershipBenefitType(ctx context.Context, req *DeleteMembershipBenefitTypeReq, opts ...http.CallOption) (rsp *DeleteMembershipBenefitTypeReply, err error)
	GetMembershipBenefitTypeInfo(ctx context.Context, req *GetMembershipBenefitTypeInfoReq, opts ...http.CallOption) (rsp *GetMembershipBenefitTypeInfoReply, err error)
	GetMembershipBenefitTypeList(ctx context.Context, req *GetMembershipBenefitTypeListReq, opts ...http.CallOption) (rsp *GetMembershipBenefitTypeListReply, err error)
	UpdateMembershipBenefitType(ctx context.Context, req *UpdateMembershipBenefitTypeReq, opts ...http.CallOption) (rsp *UpdateMembershipBenefitTypeReply, err error)
	UpdateMembershipBenefitTypeStatus(ctx context.Context, req *UpdateMembershipBenefitTypeStatusReq, opts ...http.CallOption) (rsp *UpdateMembershipBenefitTypeStatusReply, err error)
}

type MembershipBenefitTypeHTTPClientImpl struct {
	cc *http.Client
}

func NewMembershipBenefitTypeHTTPClient(client *http.Client) MembershipBenefitTypeHTTPClient {
	return &MembershipBenefitTypeHTTPClientImpl{client}
}

func (c *MembershipBenefitTypeHTTPClientImpl) CreateMembershipBenefitType(ctx context.Context, in *CreateMembershipBenefitTypeReq, opts ...http.CallOption) (*CreateMembershipBenefitTypeReply, error) {
	var out CreateMembershipBenefitTypeReply
	pattern := "/admin/v1/membership_benefit_type/create"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipBenefitTypeCreateMembershipBenefitType))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *MembershipBenefitTypeHTTPClientImpl) DeleteMembershipBenefitType(ctx context.Context, in *DeleteMembershipBenefitTypeReq, opts ...http.CallOption) (*DeleteMembershipBenefitTypeReply, error) {
	var out DeleteMembershipBenefitTypeReply
	pattern := "/admin/v1/membership_benefit_type/delete"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipBenefitTypeDeleteMembershipBenefitType))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *MembershipBenefitTypeHTTPClientImpl) GetMembershipBenefitTypeInfo(ctx context.Context, in *GetMembershipBenefitTypeInfoReq, opts ...http.CallOption) (*GetMembershipBenefitTypeInfoReply, error) {
	var out GetMembershipBenefitTypeInfoReply
	pattern := "/admin/v1/membership_benefit_type/info"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMembershipBenefitTypeGetMembershipBenefitTypeInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *MembershipBenefitTypeHTTPClientImpl) GetMembershipBenefitTypeList(ctx context.Context, in *GetMembershipBenefitTypeListReq, opts ...http.CallOption) (*GetMembershipBenefitTypeListReply, error) {
	var out GetMembershipBenefitTypeListReply
	pattern := "/admin/v1/membership_benefit_type/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMembershipBenefitTypeGetMembershipBenefitTypeList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *MembershipBenefitTypeHTTPClientImpl) UpdateMembershipBenefitType(ctx context.Context, in *UpdateMembershipBenefitTypeReq, opts ...http.CallOption) (*UpdateMembershipBenefitTypeReply, error) {
	var out UpdateMembershipBenefitTypeReply
	pattern := "/admin/v1/membership_benefit_type/update"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipBenefitTypeUpdateMembershipBenefitType))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *MembershipBenefitTypeHTTPClientImpl) UpdateMembershipBenefitTypeStatus(ctx context.Context, in *UpdateMembershipBenefitTypeStatusReq, opts ...http.CallOption) (*UpdateMembershipBenefitTypeStatusReply, error) {
	var out UpdateMembershipBenefitTypeStatusReply
	pattern := "/admin/v1/membership_benefit_type/update/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMembershipBenefitTypeUpdateMembershipBenefitTypeStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}