# ai-boilerplate-backend/doc/swagger/admin/v1/membership_benefit_type.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "admin/v1/membership_benefit_type.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MembershipBenefitType"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/admin/v1/membership_benefit_type/create": {
      "post": {
        "summary": "会员权益类型表-创建一条数据",
        "operationId": "MembershipBenefitType_CreateMembershipBenefitType",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/admin.v1.CreateMembershipBenefitTypeReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/admin.v1.CreateMembershipBenefitTypeReq"
            }
          },
          {
            "name": "Authorization",
            "description": "Bearer Token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MembershipBenefitType"
        ]
      }
    },
    "/admin/v1/membership_benefit_type/delete": {
      "post": {
        "summary": "会员权益类型表-删除一条数据",
        "operationId": "MembershipBenefitType_DeleteMembershipBenefitType",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/admin.v1.DeleteMembershipBenefitTypeReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/admin.v1.DeleteMembershipBenefitTypeReq"
            }
          },
          {
            "name": "Authorization",
            "description": "Bearer Token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MembershipBenefitType"
        ]
      }
    },
    "/admin/v1/membership_benefit_type/info": {
      "get": {
        "summary": "会员权益类型表-单条数据查询",
        "operationId": "MembershipBenefitType_GetMembershipBenefitTypeInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/admin.v1.GetMembershipBenefitTypeInfoReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "id",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "description": "Bearer Token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MembershipBenefitType"
        ]
      }
    },
    "/admin/v1/membership_benefit_type/list": {
      "get": {
        "summary": "会员权益类型表-列表数据查询",
        "operationId": "MembershipBenefitType_GetMembershipBenefitTypeList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/admin.v1.GetMembershipBenefitTypeListReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "description": "页码",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "description": "页数",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "benefitKey",
            "description": "权益标识",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "benefitName",
            "description": "权益名称",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "description": "Bearer Token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MembershipBenefitType"
        ]
      }
    },
    "/admin/v1/membership_benefit_type/update": {
      "post": {
        "summary": "会员权益类型表-更新一条数据",
        "operationId": "MembershipBenefitType_UpdateMembershipBenefitType",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/admin.v1.UpdateMembershipBenefitTypeReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/admin.v1.UpdateMembershipBenefitTypeReq"
            }
          },
          {
            "name": "Authorization",
            "description": "Bearer Token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MembershipBenefitType"
        ]
      }
    },
    "/admin/v1/membership_benefit_type/update/status": {
      "post": {
        "summary": "会员权益类型表-更新状态",
        "operationId": "MembershipBenefitType_UpdateMembershipBenefitTypeStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/admin.v1.UpdateMembershipBenefitTypeStatusReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/admin.v1.UpdateMembershipBenefitTypeStatusReq"
            }
          },
          {
            "name": "Authorization",
            "description": "Bearer Token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MembershipBenefitType"
        ]
      }
    }
  },
  "definitions": {
    "admin.v1.CreateMembershipBenefitTypeReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id"
        }
      },
      "title": "响应-会员权益类型表-创建一条数据"
    },
    "admin.v1.CreateMembershipBenefitTypeReq": {
      "type": "object",
      "properties": {
        "benefitKey": {
          "type": "string",
          "title": "权益标识"
        },
        "benefitName": {
          "type": "string",
          "title": "权益名称"
        },
        "benefitIcon": {
          "type": "string",
          "title": "权益图标"
        },
        "benefitDesc": {
          "type": "string",
          "title": "权益描述"
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "title": "排序"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "title": "状态(-1禁用,1启用)"
        }
      },
      "title": "请求-会员权益类型表-创建一条数据",
      "required": [
        "benefitKey",
        "benefitName",
        "status"
      ]
    },
    "admin.v1.DeleteMembershipBenefitTypeReply": {
      "type": "object",
      "title": "响应-会员权益类型表-删除一条数据"
    },
    "admin.v1.DeleteMembershipBenefitTypeReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id"
        }
      },
      "title": "请求-会员权益类型表-删除一条数据",
      "required": [
        "id"
      ]
    },
    "admin.v1.GetMembershipBenefitTypeInfoReply": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/admin.v1.MembershipBenefitTypeInfo"
        }
      },
      "title": "响应-会员权益类型表-单条数据查询"
    },
    "admin.v1.GetMembershipBenefitTypeListReply": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int32",
          "title": "总数"
        },
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/admin.v1.MembershipBenefitTypeInfo"
          },
          "title": "列表数据"
        }
      },
      "title": "响应-会员权益类型表-列表数据查询"
    },
    "admin.v1.MembershipBenefitTypeInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id"
        },
        "benefitKey": {
          "type": "string",
          "title": "权益标识"
        },
        "benefitName": {
          "type": "string",
          "title": "权益名称"
        },
        "benefitIcon": {
          "type": "string",
          "title": "权益图标"
        },
        "benefitDesc": {
          "type": "string",
          "title": "权益描述"
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "title": "排序"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "title": "状态(-1禁用,1启用)"
        },
        "createdAt": {
          "type": "string",
          "title": "创建时间"
        },
        "updatedAt": {
          "type": "string",
          "title": "更新时间"
        }
      },
      "title": "会员权益类型表信息"
    },
    "admin.v1.UpdateMembershipBenefitTypeReply": {
      "type": "object",
      "title": "响应-会员权益类型表-更新一条数据"
    },
    "admin.v1.UpdateMembershipBenefitTypeReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id"
        },
        "benefitKey": {
          "type": "string",
          "title": "权益标识"
        },
        "benefitName": {
          "type": "string",
          "title": "权益名称"
        },
        "benefitIcon": {
          "type": "string",
          "title": "权益图标"
        },
        "benefitDesc": {
          "type": "string",
          "title": "权益描述"
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "title": "排序"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "title": "状态(-1禁用,1启用)"
        }
      },
      "title": "请求-会员权益类型表-更新一条数据",
      "required": [
        "id",
        "benefitKey",
        "benefitName",
        "status"
      ]
    },
    "admin.v1.UpdateMembershipBenefitTypeStatusReply": {
      "type": "object",
      "title": "响应-会员权益类型表-更新状态"
    },
    "admin.v1.UpdateMembershipBenefitTypeStatusReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "title": "状态(-1禁用,1启用)"
        }
      },
      "title": "请求-会员权益类型表-更新状态",
      "required": [
        "id",
        "status"
      ]
    },
    "google.protobuf.Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "google.rpc.Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/google.protobuf.Any"
          }
        }
      }
    }
  }
}